requirements.txt
asgiref==3.5.0
Django==4.0.3
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
